requirements.txt
django==1.9
djangorestframework==3.3.3
python-social-auth==0.2.14
django-rest-framework-social-oauth2==1.0.4
django-model-utils==2.4